apps/ai-agent/pyproject.toml
[project]
name = "novyra-ai-engine"
version = "2.0.0"
description = "NOVYRA AI Engine — Structured Reasoning, Knowledge Graph, Mastery Tracking"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=23.2.0",
    "fastapi>=0.115.0",
    "google-generativeai>=0.8.0",
    "neo4j>=5.20.0",
    "gunicorn>=22.0.0",
    "httpx>=0.27.0",
    "mangum>=0.17.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pypdf>=4.3.0",
    "python-docx>=1.1.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "tenacity>=8.3.0",
    "deep-translator>=1.11.0",
    "langdetect>=1.0.9",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "uvicorn[standard]>=0.30.0",
    "requests>=2.32.0",
]
